Rework inaccuracy checker, normalisation and embedder
Previously, the inaccuracy checker's structure did not allow for
sufficient flexibility in inputting data and reporting inaccuracies.
Moreover, the normalisation process was implemented in two different
ways, with the embedders using a different method to the inaccuracy
checkers.

This reworks the inaccuracy checkers to allow for a more flexible input
and additional functionality. Furthermore, the XML extraction and CSV
normalisation processes were refactored, reworked and adapted for
properties. All other classes that used these functions have been
adapted to accommodate the reworked functionalities.

Finally, the logs from the inaccuracy checker were saved to be used in
our current paper.